+++ src/main/java/org/kumoricon/registration/BaseDataService.java
@@ -5,7 +5,6 @@ import org.kumoricon.registration.model.role.Right;
 import org.kumoricon.registration.model.role.RightRepository;
 import org.kumoricon.registration.model.role.Role;
 import org.kumoricon.registration.model.role.RoleRepository;
-import org.kumoricon.registration.model.tillsession.TillSession;
 import org.kumoricon.registration.model.tillsession.TillSessionRepository;
 import org.kumoricon.registration.model.user.User;
 import org.kumoricon.registration.model.user.UserRepository;
@@ -15,8 +14,6 @@ import org.kumoricon.registration.settings.SettingsService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-
-import java.time.OffsetDateTime;
 import java.util.*;
 
 
@@ -27,7 +24,6 @@ public class BaseDataService {
     private final UserRepository userRepository;
     private final UserService userService;
     private final BadgeService badgeService;
-    private final TillSessionRepository tillSessionRepository;
     private final Settings currentSettings;
     private static final Logger log = LoggerFactory.getLogger(BaseDataService.class);
 
@@ -36,14 +32,12 @@ public class BaseDataService {
                            UserRepository userRepository,
                            UserService userService,
                            BadgeService badgeService,
-                           TillSessionRepository tillSessionRepository,
                            SettingsService settingsService) {
         this.roleRepository = roleRepository;
         this.rightRepository = rightRepository;
         this.userRepository = userRepository;
         this.userService = userService;
         this.badgeService = badgeService;
-        this.tillSessionRepository = tillSessionRepository;
         this.currentSettings = settingsService.getCurrentSettings();
     }
 
@@ -160,6 +154,7 @@ public class BaseDataService {
                 {"badge_type_panelist", "Select/check in the \"Panelist\" badge type"},
                 {"badge_type_staff", "Select/check in the \"Staff\" badge type"},
                 {"view_attendance_report", "View attendance report"},
+                {"view_attendance_report_deduplicated", "View experimental deduplicated attendance report"},
                 {"view_check_in_by_badge_type_report", "View check in by badge type report (counts only)"},
                 {"view_check_in_by_hour_report", "View attendee check ins per hour report"},
                 {"view_check_in_by_user_report", "View attendee check ins per user report"},